Extend < and > word forms; escape colliding tags
Extend words that begin with < or > to allow any number of characters
from - = < | > + ~ after the initial angle character (angle-words).

Introduce the period '.' character as a tag-escape marker to cause tag
interpretation when tags and words overlap; e.g., make "<. >" load as
the tag that has as its contents a single blank character.

Preserve old behaviour when the tag contents begin with a / character.

Similar change to MOLD is to be made subsequently.

Eight different circumstances require this tag/word discrimination:
1) Refinement processing (leading / character).
2) Lit-word processing (leading ' character).
3) Get-word processing (leading : character).
4) Issue processing (leading # character).
5) Angle-word processing (leading < or > character).
6) Word delimitation checking (word abuts > or < character).
7) Tag post-processing, when removing the escaping '.' character.
8) Converting strings to issues, to be valid for angle-words.

Since angle-words cannot begin URLs, Scan_Block no longer
assumes that in-path word tokens can end with colon only at path end.
Adjust word token handling to syntax error if a lit- or get-word that
ends in colon, or a set-word, occurs before the end of the path.
This assumption was erroneously implemented originally for lit and
get words, and is fully corrected (ruled out) with this code change.

Converting strings to issues needs word-type content model definitions,
so in the meantime just allow issues to have < and > in them.

Details:

Tags with contents of
    (zero or more - = < | > + ~ then)
    (zero or some periods then)
        nothing
    or
       (an optional : then)
       whitespace or ] / ) " { ; ( [ then
       anything
and that do not start directly with /
loadably mold with an extra, non-content, period.

CTSC*, or the Closing Tag Special Case:
    / appearing after the single-<-word means it is a tag,
    instead of delimiting the word and starting a path,
    like / does after any other word.

Implements:
1) Tags can begin with whitespace by escaping.
2) Tags can begin with right-bracket by escaping.
3) Tags can begin with <, =, or > then (colon and) space by escaping.
4) Tags can be empty by escaping.
5) Other words that begin with < or > are available.
6) Words that begin with < or > are delimited by close-paren ).
7) Words that begin with < or > are delimited by open-string " or {.
8) Words that begin with < or > are delimited by open-block [.
9) Words that begin with < or > are delimited by open-paren (.
10) Words that begin with < or > are delimited by open-comment ;.
11) Words that begin with < or > are available as set-words.
12) Words that begin with < or > excepting < are able to begin paths.
13) Words that begin with < or > are available as issues.
14) Words that begin with < or > are available as refinements.

Fixes:
Allow constructs, strings, and blocks in paths.
Allow moneys, decimals, and percents in mid-path [a/1.1/b].
Change signed and fractional times to times, not set-words [:+1].
Disallow lit-words that begin with signed or unsigned fractions ['-.2].
Disallow > in words when delimited by a following tag [a>><b>].
Syntax error 'a:b and :a:b instead of allowing colons in word types.
Syntax error 'a/b: and :a/b: instead of stripping the last colon.
Syntax error 'a/1:/b and :a/1:/b instead of stripping the last colon.
Syntax error 'a/'b:/c and 'a/:b:/c instead of stripping the last colon.
Disallow -# and +# forms.
Change error on :a/b: to "invalid path" from "invalid word-get".
Disallow conversion to word! of "a]b".
Raise error on invalid UTF-8 rather than returning "^@".

Examples of effects, or lack thereof:

Load String  Changed?  Before                   After
===========  ========  ======                   =====
< >          No        Word, word               Word, word
<. >         Yes       Tag of ". "              Tag of " "
<a>          No        Tag of "a"               Tag of "a"
<.a>         No        Tag of ".a"              Tag of ".a"
<;>          Yes       Tag of ";"               Word, comment
<"">         Yes       Tag of {""}              Word, string, word
<{}>         Yes       Tag of "{}"              Word, string, word
<]>          No        Word, end-block, word    Word, end-block, word
<[>          Yes       Tag of "["               Word, begin-block, word
<)>          Yes       Tag of ")"               Word, end-paren, word
<(>          Yes       Tag of "("               Word, begin-paren, word
</>          No        Tag of "/"               Tag of "/"
<: >         Yes       Tag of ": "              Set-word, word
<.: >        Yes       Tag of ".: "             Tag of ": "
<:. >        No        Tag of ":. "             Tag of ":. "
<a >         No        Tag of "a "              Tag of "a "
<:a >        No        Tag of ":a "             Tag of ":a "
<a: >        No        Tag of "a: "             Tag of "a: "
<--          Yes       Invalid tag              Word
<--a         No        Invalid tag              Invalid tag
<--:         Yes       Invalid tag              Set-word
<--:a        No        Invalid tag              Invalid tag
<--:a>       No        Tag of "--:a"            Tag of "--:a"
<--: a>      Yes       Tag of "--: a"           Set-word, invalid word
<-->         Yes       Tag of "--"              Word
<.-->        No        Tag of ".--"             Tag of ".--"
<-.->        No        Tag of "-.-"             Tag of "-.-"
<--.>        Yes       Tag of "--."             Tag of "--"
<--..>       Yes       Tag of "--.."            Tag of "--."
<-->.        Yes       Tag of "--", word        Invalid tag
<-->.>       Yes       Invalid word             Tag of "-->"
<><>         Yes       Invalid word             Word
<.><>        Yes       Tag of ".", word of <>   Tag of "", word of <>
<>.<>        Yes       Invalid word             Tag of ">.<"
<><.>        Yes       Invalid word             Tag of "><"
</a>         No        Tag of "/a"              Tag of "/a"
<./a>        Yes       Tag of "./a"             Tag of "/a"
<../a>       Yes       Tag of "../a"            Tag of "./a"

Finally, ([a]<[b])=([b]>[a]) now evaluates to true instead of to [a].